requirements.txt
moviepy
openai
python-dotenv
